+++ src/datasources/data-frame/components/v2/DataFrameQueryEditorV2.test.tsx
@@ -12,7 +12,7 @@ import { DataFrameDataSource } from "datasources/data-frame/DataFrameDataSource"
 import { DataFrameQueryBuilderWrapper } from "./query-builders/DataFrameQueryBuilderWrapper";
 import { COLUMN_OPTIONS_LIMIT } from "datasources/data-frame/constants";
 import { ComboboxOption } from "@grafana/ui";
-import { errorMessages } from "datasources/data-frame/constants/v2/DataFrameQueryEditorV2.constants";
+import { errorMessages, infoMessages } from "datasources/data-frame/constants/v2/DataFrameQueryEditorV2.constants";
 import { of } from "rxjs";
 
 jest.mock("./query-builders/DataFrameQueryBuilderWrapper", () => ({
@@ -296,18 +296,6 @@ describe("DataFrameQueryEditorV2", () => {
                 });
 
                 describe('column limit', () => {
-                    it('should not render warning alert when column limit is not exceeded', async () => {
-                        await changeFilterValue();
-                        await clickColumnOptions();
-
-                        const optionTexts = getColumnOptionTexts();
-
-                        expect(optionTexts.length).toBeLessThanOrEqual(COLUMN_OPTIONS_LIMIT);
-                        await waitFor(() => {
-                            expect(screen.queryByRole('alert')).not.toBeInTheDocument();
-                        });
-                    });
-
                     describe('when column limit is exceeded', () => {
                         beforeEach(async () => {
                             cleanup();
@@ -330,41 +318,6 @@ describe("DataFrameQueryEditorV2", () => {
                             await clickColumnOptions();
                             const optionTexts = getColumnOptionTexts();
                             expect(optionTexts.length).toBe(COLUMN_OPTIONS_LIMIT);
-                        });
-
-                        it('should render warning alert when column limit is exceeded', async () => {
-                            await waitFor(() => {
-                                const alert = screen.getByRole('alert');
-                                expect(alert).toBeInTheDocument();
-                                expect(within(alert).getByText(/Warning/i)).toBeInTheDocument();
-                                expect(within(alert).getByText(errorMessages.columnLimitExceeded)).toBeInTheDocument();
-                            });
-                        });
-
-                        it('should hide the warning alert when filter is removed', async () => {
-                            await changeFilterValue();
-                            const alert = screen.getByRole('alert');
-                            expect(alert).toBeInTheDocument();
-
-                            await changeFilterValue('');
-
-                            await waitFor(() => {
-                                expect(alert).not.toBeInTheDocument();
-                            });
-                        });
-
-                        it('should hide the warning alert when filter is changed and the `getColumnOptionsWithVariables` errored', async () => {
-                            await changeFilterValue();
-                            const alert = screen.getByRole('alert');
-                            expect(alert).toBeInTheDocument();
-
-                            // Mock getColumnOptionsWithVariables to throw error
-                            jest.spyOn(datasource, 'getColumnOptionsWithVariables').mockRejectedValue(new Error('Test error'));
-                            await changeFilterValue('AnotherFilter');
-
-                            await waitFor(() => {
-                                expect(alert).not.toBeInTheDocument();
-                            });
                         });
                     });
                 });
@@ -1660,18 +1613,6 @@ describe("DataFrameQueryEditorV2", () => {
                 });
 
                 describe('x-column limit', () => {
-                    it('should not render warning alert when x-column limit is not exceeded', async () => {
-                        await changeFilterValue();
-                        await clickXColumnCombobox();
-
-                        const optionTexts = getXColumnOptionTexts();
-
-                        expect(optionTexts.length).toBeLessThanOrEqual(COLUMN_OPTIONS_LIMIT);
-                        await waitFor(() => {
-                            expect(screen.queryByRole('alert')).not.toBeInTheDocument();
-                        });
-                    });
-
                     describe('when x-column limit is exceeded', () => {
                         beforeEach(async () => {
                             cleanup();
@@ -1697,42 +1638,6 @@ describe("DataFrameQueryEditorV2", () => {
                             await clickXColumnCombobox();
                             const optionTexts = getXColumnOptionTexts();
                             expect(optionTexts.length).toBe(COLUMN_OPTIONS_LIMIT);
-                        });
-
-                        it('should render warning alert when x-column limit is exceeded', async () => {
-                            await waitFor(() => {
-                                const alert = screen.getByRole('alert');
-                                expect(alert).toBeInTheDocument();
-                                expect(within(alert).getByText(/Warning/i)).toBeInTheDocument();
-                                expect(within(alert).getByText(errorMessages.xColumnLimitExceeded))
-                                    .toBeInTheDocument();
-                            });
-                        });
-
-                        it('should hide the warning alert when filter is removed', async () => {
-                            await changeFilterValue();
-                            const alert = screen.getByRole('alert');
-                            expect(alert).toBeInTheDocument();
-
-                            await changeFilterValue('');
-
-                            await waitFor(() => {
-                                expect(alert).not.toBeInTheDocument();
-                            });
-                        });
-
-                        it('should hide the warning alert when filter is changed and the `getColumnOptionsWithVariables` errored', async () => {
-                            await changeFilterValue();
-                            const alert = screen.getByRole('alert');
-                            expect(alert).toBeInTheDocument();
-
-                            // Mock getColumnOptionsWithVariables to throw error
-                            jest.spyOn(datasource, 'getColumnOptionsWithVariables').mockRejectedValue(new Error('Test error'));
-                            await changeFilterValue('AnotherFilter');
-
-                            await waitFor(() => {
-                                expect(alert).not.toBeInTheDocument();
-                            });
                         });
                     });
                 });
@@ -2927,6 +2832,14 @@ describe("DataFrameQueryEditorV2", () => {
             const [[props]] = (DataFrameQueryBuilderWrapper as jest.Mock).mock.calls;
 
             expect(props.dataTableFilter).toBe('TestFilter');
+        });
+
+        it('should pass datasourceHelp message as additionalInfoMessage to the wrapper', () => {
+          renderComponent({ type: DataFrameQueryType.Data, dataTableFilter: 'TestFilter' });
+
+          const [[props]] = (DataFrameQueryBuilderWrapper as jest.Mock).mock.calls;
+
+          expect(props.additionalInfoMessage).toBe(infoMessages.datasourceHelp);
         });
 
         it("should pass columnFilter to the query builder wrapper", () => {