+++ src/datasources/data-frame/components/v2/DataFrameQueryEditorV2.tsx
@@ -14,6 +14,7 @@ import {
     labels,
     placeholders,
     tooltips,
+    infoMessages,
 } from 'datasources/data-frame/constants/v2/DataFrameQueryEditorV2.constants';
 import { isObservable, lastValueFrom } from 'rxjs';
 import _ from 'lodash';
@@ -26,10 +27,8 @@ export const DataFrameQueryEditorV2: React.FC<Props> = ({ query, onChange, onRun
     const [isDecimationSettingsSectionOpen, setIsDecimationSettingsSectionOpen] = useState(true);
     const [recordCountInvalidMessage, setRecordCountInvalidMessage] = useState<string>('');
     const [columnOptions, setColumnOptions] = useState<Array<ComboboxOption<string>>>([]);
-    const [isColumnLimitExceeded, setIsColumnLimitExceeded] = useState<boolean>(false);
     const [isPropertiesNotSelected, setIsPropertiesNotSelected] = useState<boolean>(false);
     const [xColumnOptions, setXColumnOptions] = useState<Array<ComboboxOption<string>>>([]);
-    const [isXColumnLimitExceeded, setIsXColumnLimitExceeded] = useState<boolean>(false);
     const [isColumnOptionsInitialized, setIsColumnOptionsInitialized] = useState<boolean>(false);
 
     const getPropertiesOptions = (
@@ -63,19 +62,11 @@ export const DataFrameQueryEditorV2: React.FC<Props> = ({ query, onChange, onRun
                     .slice(0, COLUMN_OPTIONS_LIMIT);
                 const limitedXColumnOptions = columnOptions.commonColumnsAcrossTables
                     .slice(0, COLUMN_OPTIONS_LIMIT);
-                setIsColumnLimitExceeded(
-                    columnOptions.uniqueColumnsAcrossTables.length > COLUMN_OPTIONS_LIMIT
-                );
-                setIsXColumnLimitExceeded(
-                    columnOptions.commonColumnsAcrossTables.length > COLUMN_OPTIONS_LIMIT
-                );
                 setColumnOptions(limitedColumnOptions);
                 setXColumnOptions(limitedXColumnOptions);
             } catch (error) {
                 setColumnOptions([]);
                 setXColumnOptions([]);
-                setIsColumnLimitExceeded(false);
-                setIsXColumnLimitExceeded(false);
             } finally {
                 setIsColumnOptionsInitialized(true);
             }
@@ -168,8 +159,6 @@ export const DataFrameQueryEditorV2: React.FC<Props> = ({ query, onChange, onRun
             }
 
             // Clear column options if filter is empty
-            setIsColumnLimitExceeded(false);
-            setIsXColumnLimitExceeded(false);
             if (columnOptions.length > 0) {
                 setColumnOptions([]);
             }
@@ -372,21 +361,12 @@ export const DataFrameQueryEditorV2: React.FC<Props> = ({ query, onChange, onRun
                     collapsible={true}
                     onToggle={() => setIsQueryConfigurationSectionOpen(!isQueryConfigurationSectionOpen)}
                 >
-                    {migratedQuery.type === DataFrameQueryType.Data && (
-                        <>
-                            {isColumnLimitExceeded && (
-                                <Alert title='Warning' severity='warning'>{errorMessages.columnLimitExceeded}</Alert>
-                            )}
-                            {isXColumnLimitExceeded && (
-                                <Alert title='Warning' severity='warning'>{errorMessages.xColumnLimitExceeded}</Alert>
-                            )}
-                        </>
-                    )}
                     <DataFrameQueryBuilderWrapper
                         datasource={datasource}
                         resultFilter={migratedQuery.resultFilter}
                         dataTableFilter={migratedQuery.dataTableFilter}
                         columnFilter={migratedQuery.columnFilter}
+                        additionalInfoMessage={infoMessages.datasourceHelp}
                         onResultFilterChange={onResultFilterChange}
                         onDataTableFilterChange={onDataTableFilterChange}
                         onColumnFilterChange={onColumnFilterChange}
@@ -520,7 +500,11 @@ export const DataFrameQueryEditorV2: React.FC<Props> = ({ query, onChange, onRun
             {migratedQuery.type === DataFrameQueryType.Properties && (
                 <>
                     {isPropertiesNotSelected && (
-                        <Alert title='Error' severity='error'>
+                        <Alert 
+                            title='Error' 
+                            severity='error' 
+                            style={{ width: getValuesInPixels(VALUE_FIELD_WIDTH) }}
+                        >
                             {errorMessages.propertiesNotSelected}
                         </Alert>
                     )}
@@ -537,6 +521,7 @@ export const DataFrameQueryEditorV2: React.FC<Props> = ({ query, onChange, onRun
                             value={migratedQuery.dataTableProperties}
                             onChange={onDataTablePropertiesChange}
                             options={dataTablePropertiesOptions}
+                            isClearable={true}
                         />
                     </InlineField>
                     <InlineField
@@ -552,6 +537,7 @@ export const DataFrameQueryEditorV2: React.FC<Props> = ({ query, onChange, onRun
                             value={migratedQuery.columnProperties}
                             onChange={onColumnPropertiesChange}
                             options={columnPropertiesOptions}
+                            isClearable={true}
                         />
                     </InlineField>
                 </>